+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `pcb search` now supports `kicad:components` in both the TUI and non-interactive search.
+
 ## [0.3.58] - 2026-03-21
 
 ### Added
@@ -15,7 +19,6 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 - `Usb2TypeC` interface adapter in stdlib to connet USB2 to TypeC
 - `Dvp` interface for cameras with 8, 10, 12 and 16 bit width
 - `SdRam` interface for 16 bit and 32 bit
-- `pcb search` now supports `kicad:components` in both the TUI and non-interactive search.
 
 ### Changed